requirements.txt
pyqt5-tools~=5.15
